fix(logging): let the cli configure logging
This allows for a user using the library to configure logging, e.g. use
`logging.getLogger('dpat').propagate = False`, if logging is not needed.
The cli will always log to the terminal.

BREAKING CHANGE: logging with the config file is now unsupported.
Configure logging in the application using the library.