feat(installation): remove the need for config.yml
To install dpat in a script to use as api on windows, it is able
to achieve that with
import dpat
dpat.install_windows("path/to/vips/bin")

BREAKING CHANGE: installation via the config.yml is no longer possible.
It is also no longer needed for splits/convert cli operations.
For coming deep learning cli applications, it will be needed to fetch
the path to vipsbin from a config with deep learning options.