fix: filter h5 dataset by filenames
Using the paths_and_targets keyword, which is equivalent to the similar
similar keyword in the tile dataset.

BREAKING CHANGE: train|val|test_path keywords now do not refer to the
the target train|val|test paths to output h5 files to, but to the files
providing paths_and_targets.
file_path is the new target file, containing all images.